% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R89P3S09}
\alias{R89P3S09}
\title{Table R89P3S09 \cr
Household recreation & leisure expenditure of rural household in 1389 \cr
هزينه هاي  مربوط به خدمات فرهنگي و تفريحات خانوار روستایی سال 1389}
\format{
A data frame with 4 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{ارزش (ریال)}
\item Description: Value (Rials)
}
}
}
\usage{
R89P3S09
}
\description{
{R for Rural} \cr
{89 for 1389} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
